moved charts and features calculation into own classes
